Move ApiController inside the API module
